requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.13.0
backcall==0.2.0
bangla==0.0.2
blinker==1.6.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1696431134479/work
ci-sdr==0.0.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.1.1
coqpit==0.0.17
ctc-segmentation==1.7.4
cycler==0.12.1
Cython==0.29.30
dateparser==1.1.8
decorator==5.1.1
deepl==1.15.0
Distance==0.1.3
distlib==0.3.4
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.2
einops==0.6.1
emoji==2.8.0
encodec==0.1.1
espeakng==1.0.2
espnet==202301
espnet-tts-frontend==0.0.3
exceptiongroup==1.1.3
executing==2.0.0
fast-bss-eval==0.1.3
ffmpeg-python==0.2.0
filelock==3.6.0
Flask==2.3.3
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2p-en==2.1.0
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h5py==3.10.0
huggingface-hub==0.18.0
humanfriendly==10.0
hydra-core==1.3.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.31.5
importlib-metadata==4.13.0
inflect==5.6.2
ipython==8.16.1
itsdangerous==2.1.2
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
k-diffusion==0.0.16
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.0
llvmlite==0.40.1
marisa-trie==1.1.0
Markdown==3.5
MarkupSafe @ file:///D:/bld/markupsafe_1695367528298/work
matplotlib==3.7.3
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
msgpack==1.0.7
multidict==6.0.4
networkx==2.8.8
nltk==3.8.1
num2words @ git+https://github.com/savoirfairelinux/num2words.git@3e39091d052829fc9e65c18176ce7b7ff6169772
numba==0.57.0
numpy==1.22.0
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20230918
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1697423756192/work
platformdirs==2.5.2
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pynndescent==0.5.10
pyparsing==3.1.1
pypinyin==0.44.0
pyreadline3==3.4.1
pysbd==0.3.4
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytorch-wpe==0.0.1
pytz==2023.3.post1
pyworld==0.3.4
PyYAML @ file:///D:/bld/pyyaml_1695373629531/work
referencing==0.30.2
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
resize-right==0.0.2
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
stable-ts==2.13.1
stack-data==0.6.3
stanza==1.5.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tensorboard==2.15.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.1.0
torch-complex==0.4.3
torchaudio==2.1.0
torchdiffeq==0.2.3
torchsde==0.2.6
torchvision==0.16.0
tqdm==4.64.1
trainer==0.0.31
traitlets==5.11.2
trampoline==0.1.2
transformers==4.33.3
TTS==0.18.2
typeguard==2.13.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
tzlocal==5.2
ukrainian-accentor @ git+https://github.com/egorsmkv/ukrainian-accentor.git@5b7971c4e135e3ff3283336962e63fc0b1c80f4c
ukrainian-tts @ git+https://github.com/robinhad/ukrainian-tts.git@0d2509c13947de99a67739035104fcd2c72940e7
ukrainian-word-stress==1.0.2
umap-learn==0.5.4
Unidecode==1.3.7
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1697720414277/work
virtualenv==20.14.1
wandb==0.15.12
wcwidth==0.2.8
Werkzeug==3.0.0
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
yarl==1.9.2
zipp==3.17.0
